Add files via upload
Clustering Methods including: KMeans, Hierarchical, GMM, Spectrial, Birch and Kmed.
Kmed.py is the python file of implementing Kmed method, and other methods are implemented in main.py. 
In main.py, the test data is approximately normally distributed in centers (-5,-5,-5), (0,0,0) and (5,5,5).
The main.py will print the class list of the data.